Refactored router code and changed the README file. #9
