+++ README.md
@@ -1,2 +1,7 @@
 # Sentimentalytics
 A web application for statisticians who want to see a sentiment analysis on certain words or phrases.
+
+## Demo
+(Mashape Sentiment API is broken. Currently working to find another sentiment analysis API source.)
+![sentimentalytics-issue-3](https://user-images.githubusercontent.com/16450416/31626063-2945fff2-b25d-11e7-8b26-45000586f975.gif)
+![sentimentalytics-issue-9](https://user-images.githubusercontent.com/16450416/31687312-bec9bb5a-b33d-11e7-8187-406e45d45899.gif)
