+++ api/create-api.js
@@ -1,9 +1,9 @@
 const express = require('express')
-const specialRouter = require('./router')
+const specialRouter = require('./special-router')
 const { publicDir } = require('./middlewares')
 const errorHandler = require('./error-handler')
 
-module.exports = terms => {
+module.exports = (terms, users) => {
   return express()
     .use(publicDir)
     .use('/api', specialRouter(terms))
